消除跟随人物时，灵气条跳动

+++ Code/W_Content_SpectateUnit.cs
@@ -59,9 +59,9 @@ namespace Cultivation_Way.Content
         internal static void updateStats(SpectateUnit spectate_unit)
         {
             CW_Actor cw_actor = (CW_Actor)MoveCamera.focusUnit;
-            wakan_bar.setBar(cw_actor.cw_status.wakan, cw_actor.cw_cur_stats.wakan, "/" + cw_actor.cw_cur_stats.wakan);
+            wakan_bar.setBar((int)cw_actor.cw_status.wakan, cw_actor.cw_cur_stats.wakan, "/" + cw_actor.cw_cur_stats.wakan);
             wakan_bar.transform.Find("Mask/Bar").GetComponent<Image>().color = Utils.CW_Utils_Others.get_wakan_color(cw_actor.cw_status.wakan_level, cw_actor.cw_cur_stats.wakan);
-            shield_bar.setBar(cw_actor.cw_status.shield, cw_actor.cw_cur_stats.shield, "/" + cw_actor.cw_cur_stats.shield);
+            shield_bar.setBar((int)cw_actor.cw_status.shield, cw_actor.cw_cur_stats.shield, "/" + cw_actor.cw_cur_stats.shield);
         }
     }
 }
